collector/requirements.txt
telethon
python-dotenv
requests
boto3
tweepy
instagrapi